refactor(/reservation): 패키지 구조를 도메인형으로 변경
