+++ src/main/java/roomescape/reservation/domain/ReservationDate.java
@@ -1,4 +1,4 @@
-package roomescape.domain;
+package roomescape.reservation.domain;
 
 import java.time.LocalDate;
