+++ src/main/java/roomescape/reservation/domain/ReservationName.java
@@ -1,4 +1,4 @@
-package roomescape.domain;
+package roomescape.reservation.domain;
 
 public class ReservationName {
     private final String value;
