+++ src/main/java/roomescape/reservation/repository/ReservationRepository.java
@@ -1,7 +1,7 @@
-package roomescape.repository;
+package roomescape.reservation.repository;
 
 import java.util.List;
-import roomescape.domain.Reservation;
+import roomescape.reservation.domain.Reservation;
 
 public interface ReservationRepository {
