+++ src/test/java/roomescape/util/CustomDateTimeFormatterTest.java
@@ -6,8 +6,8 @@ import java.time.LocalDate;
 import java.time.LocalTime;
 import org.junit.jupiter.api.DisplayName;
 import org.junit.jupiter.api.Test;
-import roomescape.domain.ReservationDate;
-import roomescape.domain.ReservationTime;
+import roomescape.reservation.domain.ReservationDate;
+import roomescape.reservation.domain.ReservationTime;
 
 class CustomDateTimeFormatterTest {
